Update ControllableHero.hx
Thanks for the feedback, I changed the makeGraphic function to use easily digestible variables. For the curly brace block, I went ahead with the style you recommended. I don't really like that style because it causes confusion because I keep think the block is larger than it actually is. I understand this is just my reading comprehension needing some work, so I'll try to use this style from here on out so I get used to it.